% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IbdCharacterization.R
\docType{package}
\name{IbdCharacterization-package}
\alias{IbdCharacterization}
\alias{IbdCharacterization-package}
\title{IbdCharacterization: Characterization of Inflammatory Bowel Disease Patient Cohorts}
\description{
IbdCharacterization describes the baseline demographic and clinical characteristics, as well as the occurrence of treatments and outcomes of individuals diagnosed with inflammatory bowel disease (IBD) and, specifically, with Crohn's disease and ulcerative colitis.
}
\author{
\strong{Maintainer}: Chen Yanover \email{chen@kinstitute.org.il}

}
\keyword{internal}
